model-server/requirements.txt
flask==3.0.3
tensorflow-cpu==2.20.0
Pillow==10.4.0
requests==2.32.3
numpy==1.26.4